+++ core/src/test/java/cucumber/runtime/UtilsTest.java
@@ -27,30 +27,4 @@ public class UtilsTest {
 
     public static class StaticInnerClass {
     }
-
-    @Test
-    public void doesnt_time_out_if_it_takes_too_long() throws Throwable {
-        Slow slow = new Slow();
-        Object what = Utils.invoke(slow, Slow.class.getMethod("slow"), 50);
-        assertEquals("slow", what);
-    }
-
-    @Test(expected = TimeoutException.class)
-    public void times_out_if_it_takes_too_long() throws Throwable {
-        Slow slow = new Slow();
-        Utils.invoke(slow, Slow.class.getMethod("slower"), 50);
-        fail();
-    }
-
-    public static class Slow {
-        public String slow() throws InterruptedException {
-            sleep(10);
-            return "slow";
-        }
-
-        public String slower() throws InterruptedException {
-            sleep(100);
-            return "slower";
-        }
-    }
 }