+++ groovy/src/main/java/cucumber/runtime/groovy/GroovyStepDefinition.java
@@ -3,6 +3,7 @@ package cucumber.runtime.groovy;
 import cucumber.runtime.JdkPatternArgumentMatcher;
 import cucumber.runtime.ParameterType;
 import cucumber.runtime.StepDefinition;
+import cucumber.runtime.Timeout;
 import gherkin.I18n;
 import gherkin.formatter.Argument;
 import gherkin.formatter.model.Step;
@@ -13,14 +14,16 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 public class GroovyStepDefinition implements StepDefinition {
+    private final Pattern pattern;
     private final JdkPatternArgumentMatcher argumentMatcher;
+    private final int timeoutMillis;
     private final Closure body;
     private final StackTraceElement location;
-    private final Pattern pattern;
     private GroovyBackend backend;
 
-    public GroovyStepDefinition(Pattern pattern, Closure body, StackTraceElement location, GroovyBackend backend) {
+    public GroovyStepDefinition(Pattern pattern, int timeoutMillis, Closure body, StackTraceElement location, GroovyBackend backend) {
         this.pattern = pattern;
+        this.timeoutMillis = timeoutMillis;
         this.backend = backend;
         this.argumentMatcher = new JdkPatternArgumentMatcher(pattern);
         this.body = body;
@@ -44,8 +47,14 @@ public class GroovyStepDefinition implements StepDefinition {
         return result;
     }
 
-    public void execute(I18n i18n, Object[] args) throws Throwable {
-        backend.invoke(body, args);
+    public void execute(I18n i18n, final Object[] args) throws Throwable {
+        Timeout.timeout(new Timeout.Callback<Object>() {
+            @Override
+            public Object call() throws Throwable {
+                backend.invoke(body, args);
+                return null;
+            }
+        }, timeoutMillis);
     }
 
     public boolean isDefinedAt(StackTraceElement stackTraceElement) {