+++ groovy/src/main/java/cucumber/runtime/groovy/Hooks.java
@@ -3,30 +3,42 @@ package cucumber.runtime.groovy;
 import gherkin.TagExpression;
 import groovy.lang.Closure;
 
-import static java.util.Arrays.asList;
+import java.util.ArrayList;
+import java.util.List;
 
 public class Hooks {
     public static void World(Closure body) throws Throwable {
         GroovyBackend.instance.registerWorld(body);
     }
 
-    public static void Before(Object... tagsExpressionsAndBody) throws Throwable {
-        addHook(tagsExpressionsAndBody, true);
+    public static void Before(Object... args) throws Throwable {
+        addHook(args, true);
     }
 
-    public static void After(Object... tagsExpressionsAndBody) throws Throwable {
-        addHook(tagsExpressionsAndBody, false);
+    public static void After(Object... args) throws Throwable {
+        addHook(args, false);
     }
 
     private static void addHook(Object[] tagsExpressionsAndBody, boolean before) {
-        String[] tagExpressions = new String[tagsExpressionsAndBody.length - 1];
-        System.arraycopy(tagsExpressionsAndBody, 0, tagExpressions, 0, tagExpressions.length);
-        TagExpression tagExpression = new TagExpression(asList(tagExpressions));
-        Closure body = (Closure) tagsExpressionsAndBody[tagsExpressionsAndBody.length - 1];
+        List<String> tagExpressions = new ArrayList<String>();
+        int timeoutMillis = 0;
+        Closure body = null;
+
+        for (Object o : tagsExpressionsAndBody) {
+            if (o instanceof String) {
+                tagExpressions.add((String) o);
+            } else if (o instanceof Integer) {
+                timeoutMillis = (Integer) o;
+            } else if (o instanceof Closure) {
+                body = (Closure) o;
+            }
+        }
+
+        TagExpression tagExpression = new TagExpression(tagExpressions);
         if (before) {
-            GroovyBackend.instance.addBeforeHook(tagExpression, body);
+            GroovyBackend.instance.addBeforeHook(tagExpression, timeoutMillis, body);
         } else {
-            GroovyBackend.instance.addAfterHook(tagExpression, body);
+            GroovyBackend.instance.addAfterHook(tagExpression, timeoutMillis, body);
         }
     }
 }