+++ java/src/main/java/cucumber/runtime/java/JavaHookDefinition.java
@@ -16,12 +16,14 @@ import static java.util.Arrays.asList;
 public class JavaHookDefinition implements HookDefinition {
 
     private final Method method;
+    private final int timeout;
     private final TagExpression tagExpression;
     private final int order;
     private final ObjectFactory objectFactory;
 
-    public JavaHookDefinition(Method method, String[] tagExpressions, int order, ObjectFactory objectFactory) {
+    public JavaHookDefinition(Method method, String[] tagExpressions, int order, int timeout, ObjectFactory objectFactory) {
         this.method = method;
+        this.timeout = timeout;
         tagExpression = new TagExpression(asList(tagExpressions));
         this.order = order;
         this.objectFactory = objectFactory;
@@ -54,7 +56,7 @@ public class JavaHookDefinition implements HookDefinition {
                 throw new CucumberException("Hooks must declare 0 or 1 arguments. " + method.toString());
         }
 
-        Utils.invoke(objectFactory.getInstance(method.getDeclaringClass()), method, 0, args);
+        Utils.invoke(objectFactory.getInstance(method.getDeclaringClass()), method, timeout, args);
     }
 
     @Override